ZA: Move the speech related classes in views.py to their own file
