+++ pombola/south_africa/views/__init__.py
@@ -4,4 +4,5 @@ from .home import *
 from .organisations import *
 from .person import *
 from .search import *
+from .speechviews import *
 from .views import *
